Move Franca and generic translator to input_filters [newstructure]
