+++ input_filters/franca/franca_to_ifex.py
@@ -13,13 +13,13 @@ for p in ['pyfranca', 'pyfranca/pyfranca']:
         sys.path.append(os.path.join(mydir,p))
 
 import models.ifex.ifex_ast as ifex
-import other.franca.pyfranca.pyfranca as pyfranca
-import other.franca.rule_translator as m2m
+import input_filters.franca.pyfranca.pyfranca as pyfranca
+import transformers.rule_translator as m2m
 import pyfranca.ast as franca
 import re
 
 from models.ifex.ifex_ast_construction import add_constructors_to_ifex_ast_model, ifex_ast_as_yaml
-from other.franca.rule_translator import Preparation, Constant, Unsupported
+from transformers.rule_translator import Preparation, Constant, Unsupported
 
 def translate_type_name(francaitem):
     return translate_type(francaitem)
